refactor(gh): Moved SpeckleUpgradeObject to it's own file
