+++ ConnectorGrasshopper/ConnectorGrasshopperShared/ConnectorGrasshopperShared.projitems
@@ -35,6 +35,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Extras\SpeckleStatefulParam.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extras\SpeckleStateTag.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extras\SpeckleStreamParam.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Extras\SpeckleUpgradeObject.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extras\TreeBuilder.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extras\Utilities.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Loader.cs" />
@@ -76,7 +77,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Streams\StreamGetWithTokenComponent.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Streams\StreamListComponentV2.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Streams\StreamUpdateComponentV2.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Streams\Upgraders\StreamCreateV2Upgrader.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Streams\Upgraders\V2Upgraders.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Transports\SendReceiveTransport.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Transports\Transports.Disk.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Transports\Transports.Memory.cs" />
